Refactored multicast group creation according to PR review feedback
